chore: hide action buttons after acting upon them
